test-engine/examples/api-cases_pytest/test_DS_02.py
# test_DS_02.py
# !/usr/bin/env python
# -*- coding: utf-8 -*-
import allure
import pytest
from testengine_api.core.globalContext import g_context
from testengine_api.extend.keywords import Keywords
from testengine_api.utils.VarRender import refresh

from conftest import keyWords, process_request_data


@allure.title("DS_02 提交订单-【线下支付】验证能正确的下单个商品 (不添加购物车，直接购买提交订单)")
@pytest.mark.asyncio
async def test_02():
    with allure.step("不添加购物车，直接购买"):
        url = "http://shop-xo.hctestedu.com/index.php?s=api/buy/add"
        data = {
            "buy_type": "goods",  # 类型（cart购物车、goods直接购买），不需要添加购物车
            "goods_id": "10",
            "stock": "1",
            "spec": [],
            "address_id": 921,
            "payment_id": "2",
            "site_model": 0,
            "is_points": 0,
            "user_note": ""
        }
        params = {
            "application": "app",
            "application_client_type": "weixin",
            "token": "{{token}}"
        }
        request_data = {"url": url, "params": params, "json": data}
        request_data = process_request_data(request_data)
        response = await keyWords.request_post_row(**request_data)
        
        # 结果检查
        # 1. 报文检查
        # 检查响应是否为有效的JSON格式
        try:
            if response and response.text:
                result = keyWords.ex_jsonData(EXVALUE="$..msg", VARNAME="result")
                assert result == "提交成功", f"期望'提交成功'，实际返回: {result}"
            else:
                # 如果响应为空或不是JSON格式，标记为跳过
                pytest.skip(f"API响应不可用或不是有效的JSON格式: {response.status_code if response else 'No response'}")
        except Exception as e:
            # 捕获JSON解析错误
            pytest.skip(f"JSON解析失败: {str(e)}，响应内容: {response.text if response and hasattr(response, 'text') else 'N/A'}")
