test-engine/requirements.txt
# Core Dependencies
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0
pytest-html>=3.2.0
pytest-xdist>=3.3.1
allure-pytest>=2.13.2
allure-combine

# FastAPI MCP Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
fastapi-mcp>=0.1.0
pydantic>=2.0.0

# HTTP Client
httpx>=0.24.1
requests>=2.31.0
websockets>=12.0

# Data Processing
jsonpath>=0.82
pyyaml>=6.0
pyyaml-include>=1.3
openpyxl>=3.1.2
deepdiff>=6.3.0
jinja2>=3.1.0

# Database
pymysql>=1.1.0
pymongo>=4.4.0

# Encryption
pycryptodome>=3.18.0

# Web Testing (Playwright)
playwright>=1.35.0
Pillow>=10.0.0

# Mobile Testing (Appium)
Appium-Python-Client>=3.2.1
selenium>=4.20.0
nest-asyncio>=1.5.0

# Performance Testing (Locust)
locust>=2.20.0
gevent>=23.0.0

# Utilities
loguru>=0.7.0
python-dotenv>=1.0.0

# Development Tools
ipython>=8.14.0
black>=23.7.0
flake8>=6.0.0

# Additional
faker>=18.13.0
